Seperate manipulating UI to different file
