+++ web-app/src/main/webapp/index.html
@@ -13,6 +13,7 @@
   <script
   src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDtnOIOa1q7Mwn8OeRkeXdNPxpbaWT2ZFw"
   type="text/javascript"></script>
+  <script type="module" src="script/manipulateUI.js"></script>
   <script type="module" src="script/loadData.js"></script>
   <script type="module" src="script/postUserData.js"></script>
   <script type="module" src="script/directions.js"></script>
