+++ web-app/src/main/webapp/script/script.js
@@ -15,8 +15,7 @@
 import { fetchAndParseJson, loadPoliceReports, fetchMarkers } from "/script/loadData.js";
 import { postUserReport } from "/script/postUserData.js"
 import { setDirections } from "/script/directions.js";
-
-let mapMarkers = [];
+import { showReportForm, hideReportForm, showAnalytics, hideAnalytics } from "/script/manipulateUI.js"
 
 window.onload = async () => {
   const geocoder = new google.maps.Geocoder();
@@ -129,58 +128,6 @@ function showMessageOnInfoWindow(message, position, map, infoWindow) {
   infoWindow.setPosition(position);
   infoWindow.setContent(`<h3>${message}</h3>`);
   infoWindow.open(map);
-}
-
-function showReportForm(map, geocoder) {
-  document.getElementById("form-container").style.display = "block";
-  hideHomeElements();
-
-  geocoder.geocode({ location: map.getCenter() }, (results, status) => {
-    if (status === "OK") {
-      if (results[0]) {
-        document.getElementById("location-input").value = results[0].formatted_address;
-      } else {
-        console.error("No results found");
-      }
-    } else {
-      console.error("Geocoder failed due to: " + status);
-    }
-  });
-}
-
-function hideReportForm() {
-  document.getElementById("form-container").style.display = "none";
-  showHomeElements();
-}
-
-function showAnalytics() {
-  document.getElementById("analytics-container").style.display = "block";
-  // TODO(ltwAshley): tap on map should close menu, making this unecessary
-  const menuElements = document.getElementsByClassName("menu");
-
-  for (const element of menuElements) {
-    element.style.display = "none";
-  }
-  hideHomeElements();
-}
-
-function hideAnalytics() {
-  document.getElementById("analytics-container").style.display = "none";
-  showHomeElements();
-}
-
-function showHomeElements() {
-  const homeElements = document.getElementsByClassName("home");
-  for (const element of homeElements) {
-    element.style.display = "block";
-  }
-}
-
-function hideHomeElements() {
-  const homeElements = document.getElementsByClassName("home");
-  for (const element of homeElements) {
-    element.style.display = "none";
-  }
 }
 
 function setLoginStatus(loginStatus) {
